backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.3
python-dotenv==1.0.0
newspaper3k==0.2.8
beautifulsoup4==4.12.2
transformers==4.35.2
torch==2.1.1
pydantic==2.5.2
python-multipart==0.0.6
requests==2.31.0
lxml_html_clean==0.1.1
numpy<2.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1
sentence-transformers>=3.0.1
scikit-learn==1.3.2
huggingface-hub<0.26.0
groq==0.4.2
